chore: try stackblitz samples
