+++ packages/website/package.json
@@ -37,6 +37,7 @@
     "@docusaurus/preset-classic": "3.1.1",
     "@mdx-js/react": "^3.0.0",
     "@monaco-editor/react": "^4.6.0",
+    "@stackblitz/sdk": "^1.11.0",
     "clsx": "^2.0.0",
     "docusaurus-lunr-search": "^3.3.2",
     "dotenv": "^16.4.5",
