+++ packages/website/src/components/OldEditor/Splitter.js
