+++ packages/website/src/components/OldEditor/monaco-ui5-types.d.ts
