+++ packages/website/src/components/OldEditor/sample/sample-tsconfig.js
